requirements.txt
absl-py==1.4.0
aiosignal==1.3.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
blinker==1.7.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
Django==4.2.7
einops==0.6.1
executing==1.2.0
-e git+ssh://git@github.com/SymbioticLab/FedScale.git@faab2832de4d8e32d39c379cc3cd7999992f8dd3#egg=fedscale
filelock==3.12.0
Flask==3.0.0
flatbuffers==23.5.26
flops-profiler==0.1.2
fonttools==4.43.1
frozenlist==1.3.3
fsspec==2023.5.0
gast==0.4.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.49.1
h5py==3.9.0
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.24.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
keras==2.13.1rc0
kiwisolver==1.4.5
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.6
model-profiler==1.1.8
mpmath==1.3.0
msgpack==1.0.5
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
ray==2.5.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.0
six==1.16.0
sklearn==0.0.post10
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0rc1
tensorflow-addons==0.20.0
tensorflow-estimator==2.13.0rc0
tensorflow-macos==2.13.0rc1
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.0.1
torchdata==0.6.1
torchprof==0.3.0
torchtext==0.15.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.35.2
typeguard==2.13.3
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
validators==0.20.0
vit-keras==0.1.2
wcwidth==0.2.6
Werkzeug==3.0.1
wrapt==1.15.0
zipp==3.15.0
